requirements.txt
# 基础依赖
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
datasets>=2.14.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0

# 模型微调
# peft>=0.6.0
# bitsandbytes>=0.41.0
# unsloth>=2024.0.0

# Web服务
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# 数据库与架构
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

# 数据处理
pandas>=2.0.0
albumentations>=1.3.0

# 工具
pyyaml>=6.0
tqdm>=4.66.0
matplotlib>=3.7.0
seaborn>=0.12.0

# 可选依赖（用于性能优化）
# xformers>=0.0.22
# flash-attn>=2.3.0

# 开发工具
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0